# api/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Provides an API for (Android) SMS Gateway",
        "title": "SMS Gateway - API",
        "contact": {
            "name": "Aleksandr Soloshenko",
            "email": "i@capcom.me"
        },
        "version": "{APP_VERSION}"
    },
    "host": "sms.capcom.me",
    "basePath": "/api",
    "paths": {
...
        "/api/v1/3rdparty/webhooks": {
            "get": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Returns list of registered webhooks",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Webhooks"
                ],
                "summary": "List webhooks",
                "responses": {
                    "200": {
                        "description": "Webhook list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Registers webhook. If webhook with same ID already exists, it will be replaced",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Webhooks"
                ],
                "summary": "Register webhook",
                "parameters": [
                    {
                        "description": "Webhook",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/3rdparty/webhooks/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiAuth": []
                    }
                ],
                "description": "Deletes webhook",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User",
                    "Webhooks"
                ],
                "summary": "Delete webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Webhook deleted",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/mobile/v1/webhooks": {
            "get": {
                "security": [
                    {
                        "MobileToken": []
                    }
                ],
                "description": "Returns list of registered webhooks for device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device",
                    "Webhooks"
                ],
                "summary": "List webhooks",
                "responses": {
                    "200": {
                        "description": "Webhook list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/smsgateway.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookDTO": {
            "type": "object",
            "required": [
                "event",
                "url"
            ],
            "properties": {
                "event": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookEvent"
                        }
                    ],
                    "example": "sms:received"
                },
                "id": {
                    "type": "string",
                    "maxLength": 36,
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/webhook"
                }
            }
        },
        "github_com_capcom6_sms-gateway_pkg_smsgateway.WebhookEvent": {
            "type": "string",
            "enum": [
                "sms:received"
            ],
            "x-enum-varnames": [
                "WebhookEventSmsReceived"
            ]
        },
...
        "smsgateway.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code",
                    "type": "integer"
                },
                "data": {
                    "description": "Error context"
                },
                "message": {
                    "description": "Error message",
                    "type": "string",
                    "example": "An error occurred"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuth": {
            "type": "basic"
        },
        "MobileToken": {
            "description": "Mobile device token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}