# internal/sms-gateway/handlers/webhooks/mobile.go
package webhooks

import (
	"fmt"

	"github.com/capcom6/sms-gateway/internal/sms-gateway/handlers/base"
	"github.com/capcom6/sms-gateway/internal/sms-gateway/models"
	"github.com/capcom6/sms-gateway/internal/sms-gateway/modules/auth"
	"github.com/capcom6/sms-gateway/internal/sms-gateway/modules/webhooks"
	_ "github.com/capcom6/sms-gateway/pkg/smsgateway"
	"github.com/gofiber/fiber/v2"
	"go.uber.org/fx"
	"go.uber.org/zap"
)

type mobileControllerParams struct {
	fx.In

	WebhooksServices *webhooks.Service

	Logger *zap.Logger
}

type MobileController struct {
	base.Handler

	webhooksSvc *webhooks.Service
}

//	@Summary		List webhooks
//	@Description	Returns list of registered webhooks for device
//	@Security		MobileToken
//	@Tags			Device, Webhooks
//	@Produce		json
//	@Success		200	{object}	[]smsgateway.WebhookDTO		"Webhook list"
//	@Failure		401	{object}	smsgateway.ErrorResponse	"Unauthorized"
//	@Failure		500	{object}	smsgateway.ErrorResponse	"Internal server error"
//	@Router			/mobile/v1/webhooks [get]
//
// List webhooks
func (h *MobileController) get(device models.Device, c *fiber.Ctx) error {
	items, err := h.webhooksSvc.Select(device.UserID)
	if err != nil {
		return fmt.Errorf("can't select webhooks: %w", err)
	}

	return c.JSON(items)
}

func (h *MobileController) Register(router fiber.Router) {
	router.Get("/", auth.WithDevice(h.get))
}

func NewMobileController(params mobileControllerParams) *MobileController {
	return &MobileController{
		Handler: base.Handler{
			Logger: params.Logger.Named("mobile"),
		},
		webhooksSvc: params.WebhooksServices,
	}
}